fix: Tab key navigates instead of selects
